Improve LM loss stability for fp16 training

+++ hw4lib/trainers/lm_trainer.py
@@ -54,12 +54,14 @@ class LMTrainer(BaseTrainer):
                 raw_preds, attn_weights = self.model(targets_shifted, lengths)
                 last_attn_weights = attn_weights
 
-                # CrossEntropyLoss expects [N, C] and [N]
-                B, T, V = raw_preds.size()
-                raw_loss = self.criterion(
-                    raw_preds.view(B * T, V),
-                    targets_golden.view(B * T)
-                )
+            # Compute the loss in full precision to avoid FP16 instability on
+            # CrossEntropy with large vocabularies (which can silently inflate
+            # perplexity when gradients underflow).
+            B, T, V = raw_preds.size()
+            raw_loss = self.criterion(
+                raw_preds.float().view(B * T, V),
+                targets_golden.view(B * T)
+            )
 
             # Calculate metrics with raw loss (DO NOT MODIFY THIS)
             batch_tokens = lengths.sum().item()
